hide Open Changes file action for non-local-git repositories (#578)
Fixes https://github.com/sourcegraph/sourcegraph/issues/6614

This is the quickest way to do this. We could refactor the SCM service to not have an active provider, but that'd be a much bigger change and is possibly planned upstream at https://github.com/Microsoft/vscode/issues/25838#issuecomment-317445357 (although deprecation will take a while).