disable explorer New File/Folder action bar & context menu items on readonly files (#577)
Adds a new file service method isWritable to query whether the resource at a given URI is writable.

Uses the isWritable API to update the enablement of explorer viewlet title actions and context menu actions. When a `file:` resource is focused, New File and New Folder are enabled. When a `repo:` or `repo+version:` resource is focused, those actions will be disabled.

It is still possible to trigger these actions via command quickopen, but that is acceptable.

Fixes https://github.com/sourcegraph/sourcegraph/issues/6604