prompt user for github token when needed
When the user tries to search for repos or add a repo as a root, they will be prompted to enter a GitHub personal access token to perform the search. Using a token with the 'repo' scope actually lets it work for private repos as well (currently this requires that you have a netrc for github.com though).

fixes https://github.com/sourcegraph/xsourcegraph/issues/711