fix #765 - file search hangs forever after adding root in new window
Issue #765 was occurring because the extension host gets reloaded (terminated and restarted) when the first root folder changes. When it was terminated, though, the `MainThreadWorkspace` did not dispose of the file system or file search providers registered by the old (now-terminated) extension host. So it was trying to call them and it would never hear back, thus the indefinite wait.

This is a bug in upstream as well.