monitor folder adding/removing internally to FoldersWorkbenchService to be resilient to ext host shutdown
If the extension host was shut down, then anyone monitoring a promise for when the SCM repository changed would receive a promise resolution early. This caused #775.

There is a robust way to monitor progress, which involves always using `monitorFolderOperation` (and removing the `TPromise.timeout` from `monitorFolderOperation`). There were many callers of `monitorFolderOperation` that handled this differently, so this commit simplifies the API for callers and removes the hacky timeout.

fixes #775