fix #778: remove workspace folder 'x' icon in explorer sometimes does not show up
An ineffective assignment caused actions2 to remain hidden if it was ever hidden. It would be hidden if that row was ever used for displaying a non-root item (such as a normal folder or a file).

Why set it to `undefined` not `'block'` like `actions1`? There is some CSS that makes it not show the `x` until hover, except that if the row is focused (using the keyboard), the `x` will show. This is for keyboard accessibility and screen readers.