fix #802 - do not run blame when unnecessary
Issue #802 was caused by a loop where when an output channel was opened, onDidChangeTextDocument listeners would be called for the change in the output channel's contents. The scm extension would try to blame the output channel, which would obviously fail because it isn't in a Git repo. But trying to do so would cause more output in the output channel, thus triggering the onDidChangeTextDocument listeners again, and so on.