fix parsing of compare specifier input: do not treat `--` as a revision
If you entered a string like `mybranch -- mydir/mydir2` in the SCM viewlet compare specifier input box, the list of committed changes was correct but the file diffs were incorrect because it would incorrectly consider `--` to be a revision. This fixes the parsing to match how `git diff` parses args better.

fixes #847