prefer nearer repo ancestors and exact repo matches for workspace folders
Suppose your `$HOME` is a Git repo, and you also have:

* `$HOME/foo`: not a git repo
* `$HOME/foo/bar`: is a git repo
* `$HOME/baz`: is a git repo

Then suppose you add `$HOME/foo` as a workspace root folder. Because it itself is not a Git repo but its ancestor dir `$HOME` is, the editor adds `$HOME` as an open repo in the source control viewlet. This is kind of weird but not terrible (and is desirable in some cases).

However, now `$HOME/foo/bar` and `$HOME/baz` will never get added because `getRepository(path)` first looks to see "are there any existing open repositories that contain `path`?" and that is true for all files in those two repos, because `$HOME` is an ancestor. This is bad because you can never add those other 2 repos!

This commit makes it so that workspace folders get priority in being considered as possible repo paths. If a workspace folder is itself a Git repo, then its repo is always added, even if it is inside an ancestor repo. Also, if a dir is itself a Git repo, then it is added (again, even if it is inside an ancestor Git repo).

fixes #844 (this was broken in this special case)
fixes #849