hide revision tag for non-root folders
It is confusing if a revision tag is shown next to a root folder that is NOT a Git repo, because then the revision tag actually operates on the folder's nearest ancestor Git repository, not on the folder itself. There are cases where this might be desirable, but I think it's more often undesirable. (E.g., My `$HOME` is a Git repo, so any folder that isn't itself a Git repository will have the revision tag of my `$HOME`.)

fixes #846