Prefer FolderCatalogProvider resources for exported workspaces (#821)
* Add resolveLocalFolder to FolderCatalogProvider

The intention of this is to allow going from a local path to a FolderCatalog (ie
somewhat the reverse of resolveFolder). This commit implements the
functionality by returning the first FolderCatalogProvider to return a
FolderCatalog. Currently the extensions are just stubs (always return null).

* Update API to instead be resolveLocalFolderResource: URI

Implementations of resolveLocalFolder always just called
resolveLocalFolderResource, so may as well let the client make that decision (in
case the client only needs the resource URI it can skip that step).

* Ensure resource is underneath the provider's root

* resolveLocalFolderResource for service-github

* Use FolderCatalogService to export a workspace

* Return multiple URIs

In future will allow the providers to return multiple URIs. But for now this
interface is needed to properly do reverse repo mapping.