show blame end-of-line decoration at start of hunk (or selection start), not end
addresses #934

I initially thought it was weird to not show the blame end-of-line decoration next to your cursor as you dragged a selection down (if the cursor was not on a hunk boundary). However, I was wrong about that being important, and @nicksnyder is 100% right in pointing out the weird behavior this causes.

HT @nicksnyder