Rely on resourceResolvers to indicate success of adding a root (#970)
Previously we relied on `getWorkspaceFolderForCatalogFolder` to tell when we are done cloning. However, it's implementation is flawed and relies on the git remotes matching the IFolder resource. This would only happen if the IFolder resource was a git clone URL. However, usually it was something like a `github://` resource (ie when using the repo quickopen). This would lead to behaviour where the `Adding foo/bar...` progress bar never finished.

I think we should move towards removing the `getWorkspaceFolderForCatalogFolder` function, or fix its behaviour. For now though, we update the `addFoldersAsWorkspaceRootFolders` to be much simpler and robust. The only downside of this is we now are missing some telemetry.