Wait for workspace to be fully loaded before mutating it (#1082)
When we open a deep link the navService calls
`addFoldersAsWorkspaceRootFolders` on the repo in the deep link. That repo URI
will eventually end up in the git extensions resource resolver. The first
thing the resource resolver does is check the open repositories to see if they
match the repo.  However, on startup the repositories may not be ready yet,
which would lead to the resource resolver not using an existing open
folder. So when opening a deep-link without sourcegraph already running, one
often would end up with a workspace being added twice.

This change adds a check to ensure the existing folders are ready before we add
the repo.