Revert "[extensions/service-github/src/main.ts] fix PRs diffing against wrong revision"
This reverts commit b7dd1321142261b2821972eaa73722ae3a854946.

Temporarily revert so that https://github.com/sourcegraph/src/issues/1095 works. It is more useful this way even though it sometimes is not the diff the user expects (if the branch has merge commits or if their base branch is out of date). This lets us demonstrate code intelligence in PRs. We will make every beta user we onboard aware of the limitation for now and will fix this soon in the code review work this iteration.