Bitbucket improvements (#1183)
* bitbucket: Only search for git repos

* bitbucket: Prefer recently updated repos for search

* search: Filter out invalid repos

I'm not sure how this is happening, but I was working under poor network
conditions and constantly would trigger undefined a folder. Lets just be
defensinve and filter them out.

* bitbucket: Mark repos as owned

We only search for repos we contribute to, unlike with github where we search
all public github repos.

* bitbucket: Use full_name rather than UUID for resource

We have already made the same change to github. This leads to more readable
resource URIs (which users do read sometimes). The other advantage is we rely on
the resource URI to do sorting (less than ideal). So before this change
bitbucket results never appeared compared to github results due to the resource
URL being a UUID rather than matching the query.

* bitbucket: Implement resolveLocalFolderResource

* bitbucket: Prefer HTTPS to clone

We do the same for github.

* bitbucket: Mention repo read perm at password prompt