git: New resource resolver (#1396)
This resource resolver matches the spec deep-link.md. It currently is purely
responsible for picking or cloning a repository to disk. The main difference
over the older resolver is smarter support for attempting to avoid a new
worktree.

Also included is a cleaner API intended for potential use inside of the Git
extension.

It is currently behind a feature flag since the implementation is not 100% solid
yet.