Do not include management viewlet in normal viewlets to show (#1504)
I'm not sure what has changed, but the management viewlet is now also included
in the pinned viewlet list. I assume this was related merging in master. This
change allows sourcegraph to startup, but I no longer see the settings cog in
the viewlet pane (but can click on it)