deep-links: Support opening github:// and bitbucket:// scheme (#1499)
This will allow us to remove encoding the users clone URL for github and
bitbucket. This commit only supports opening the links, it does not yet update
any generation of links.

A deep-link which excludes `vcs` param will now be directly passed to the
resource resolver. An example deep-link to gorilla/mux is
https://about.sourcegraph.com/open/#open?repo=github%3A%2F%2Fgithub.com%2Frepository%2Fgorilla%2Fmux