git: Retry operation on index lock error (#1605)
This copies the pattern used by repository.ts which retries operations if the
error was index locked. Certain operations in repository.ts are not protected by
the retry, and we copy that pattern. I'm not sure why, so if we still run into
the error we can protect those operations as well.