show context menu in view zone webviews
This makes it so that view zones have context menus appropriate to the thing the user clicked on.

To test, enable the view zone sample extension with `"sampleViewZone.enabled": true`.

This could probably be upstreamed, but that diff would actually be quite large since it needs a new service in a class that's instantiated by a lot of things manually (not via the instantiation service). So, let's try it out for ourselves for now.